Feature:  Added QuestionDetails
I added a QuestionDetails component.  This also required adding in a new Divider component and adding font-awesome as a dependency.

I also changed how I was loading in data.  Originally, I was loading in users and questions separately; however, decided to load them in at the same time to reduce complexity.